Refactor workspace route nesting and colocate DTOs (Vibe Kanban) (#3072)
* Refactor workspace route module nesting

* Move workspace route DTOs into owning modules
